refactor: move generate to another file
